Test: draft better randomized xGGQRCS test
Currently the randomized xGGQRCS test creates a pair of matrices with
uniformly distributed entries. The new randomized test creates the
matrices of the GSVD randomly and computes the matrix pair from it
because entry-wise random matrices are not really random with respect to
their singular values, see, e.g., the circular law or §28.2 in N. J.
Higham: "Accuracy and Stability of Numerical Algorithms" (2002).

Let (A, B) a matrix pair and let A = U1 D1 R Q*, B = U2 D2 R Q* be its
generalized singular value decomposition, where A* is the
complex-conjugate transpose of A. The new algorithm randomly
draws
* matrix dimensions,
* rank of (A*, B*),
* generalized singular values,
* spectral condition number of R from [1, 1/eps)

and computes all missing values with the aid of these inputs.