SGGQRCS: return sine, cosine values
Previously SGGQRCS returned angles instead of sine, cosine values but
these computations (while numerically stable) may completely destroy the
backward stability of the solver. Specifically, in one test* a computed
angle x near π/2 was the best single-precision approximation to the true
value but cosine(x) was nowhere near the true cosine value, thereby
making it impossible to achieve a backward stable reassembly of the
input matrices from the GSVD.

For this reason, SGGQRCS returns sine and cosine values to the caller
instead of angles.

TODO: Documentation

*Test name  xGGQRCS_test_singular_accuracy_vs_radians_accuracy