SGGQRCS: fix column order disagreement CS, QR
The first columns of the unitary factor of the QR factorization form a
basis for the range of the matrix at hand. This is also true for the CS
decomposition for the matrix whose CS values are in decreasing order.
For the other matrix, the basis for the matrix range is in the last
columns of the unitary factor.