SGGQRCS: add missing CS angles after pre-processing
The pre-processing may make the computation of some angles by the CS
decomposition superfluous. These missing angles should properly computed
for any combination of pre-processing flags starting with this commit.
The tests are still not passing because the pre-processing of B is still
broken when B has full rank because the bottom right element of B
containing the elementary reflector is overwritten by the scalar
belonging to the elementary reflector